Fix #32666: Memory Cache Limit stops working after set to 2048
Solved two issues here:

- RNA update function for cache limiter wasn't type-casting to size_t
  type, which lead to long int overflow.

- Display buffer size in color management wasn't calculated properly,
  ended up with much more extra memory usage than it's needed.

+++ source/blender/makesrna/intern/rna_userdef.c
@@ -265,7 +265,7 @@ static void rna_UserDef_audio_update(Main *bmain, Scene *UNUSED(scene), PointerR
 
 static void rna_Userdef_memcache_update(Main *UNUSED(bmain), Scene *UNUSED(scene), PointerRNA *UNUSED(ptr))
 {
-	MEM_CacheLimiter_set_maximum(U.memcachelimit * 1024 * 1024);
+	MEM_CacheLimiter_set_maximum(((size_t) U.memcachelimit) * 1024 * 1024);
 }
 
 static void rna_UserDef_weight_color_update(Main *bmain, Scene *scene, PointerRNA *ptr)
